refactor(ws): move websocket system to own folder
